requirements.txt
pandas
python-dotenv
python-telegram-bot
gspread
oauth2client
